requirements.txt
pandas
numpy
scikit-learn
tensorflow
fastapi
pydantic
streamlit
requests
pillow
uvicorn
matplotlib